refactor(desktop): single-home the definition form as AgentDefinitionDialog (Phase 1B.3a)
Mechanical re-homing: PersonaDialog.tsx moves to AgentDefinitionDialog.tsx
(component renamed to match its single-homed role), and its three consumers
(AgentDialog's definition family, AgentsView's edit/duplicate/import mount,
UserProfilePersonaDialogs) re-import. Zero mount-topology changes, zero
behavior changes: every surface keeps its own dialog state and mounts;
the B2 exclusivity insurance stays load-bearing until B3b/B3c collapse
each surface to a single dialog-state owner.

Testids, element ids, and user-facing copy are untouched — they move with
the form body, so the persona-path e2e specs run unchanged. Vocabulary
renames remain Phase 1B.4.

One extraction forced by the file-size guard, not preference: the
self-contained window file-drag-over effect moves to useWindowFileDragOver.ts
(behavior-identical, depth-counted enter/leave with disabled reset). The
rename-move landed the file at 1005 lines against the 1000 limit; splitting
the standalone effect is the root-cause fix, an exception would be a bandage.

Verified: tsc, biome (full src), file-size checks, 2061/0 unit tests,
smoke e2e -g "agent|persona" 56/56 locally.
