+++ desktop/src/features/agents/ui/agentCreateIntent.ts
@@ -15,8 +15,8 @@ export type AgentCreateIntent = "definition" | "definition_start" | "instance";
 
 /**
  * Default intent for callers that don't pass one. Un-migrated callers of
- * `usePersonaActions.handleSubmit` (PersonaDialog's duplicate path until B3)
- * must keep today's create-then-start semantics, so the default is
+ * `usePersonaActions.handleSubmit` (AgentDefinitionDialog's duplicate path
+ * until B3) must keep today's create-then-start semantics, so the default is
  * `definition_start`, never `definition`.
  */
 export function resolveCreateIntent(
